Move MemoryStorage to file
